refactor(receipt info)/ changed word spacing
- Tip is no longer moved when Tax value is larger.
- Since coupon is a double in the backend, made it as well on frontend.
- Cleaned up some redundant code (ReceiptInfoDialog - default values exist for a reason).

+++ app/src/main/java/com/codenode/budgetlens/data/UserReceipts.kt
@@ -54,7 +54,7 @@ class UserReceipts {
                                     val total = receipt.getDouble("total")
                                     val tax = receipt.getDouble("tax")
                                     val tip = receipt.getDouble("tip")
-                                    val coupon = receipt.getInt("coupon")
+                                    val coupon = receipt.getDouble("coupon")
                                     val currency = receipt.getString("currency")
                                     userReceipts.add(
                                         Receipts(
